Add cloudinary upload images

+++ app/Http/Controllers/InfoUserController.php
@@ -27,13 +27,35 @@ class InfoUserController extends Controller
             'password' => ['required', 'min:5', 'max:20'],
             'agreement' => ['accepted']
         ]);
+        
         $attributes['password'] = bcrypt($attributes['password']);
-
         session()->flash('success', 'Your account has been created.');
+
         $user = User::create($attributes);
         if (!isset($user->id)) {
             return back()->withErrors(['email' => 'Email o password invalido.']);
         }
+        if (request()->hasFile('profile_picture')) {
+            $profilePicture = ImageUploadController::upload(request()->file('profile_picture'), 'profile_pictures', $user->name);
+            $user->update(['profile_picture' => !empty($profilePicture) ? $profilePicture : $user->profile_picture]);
+        }
+
+        // Eliminar roles previos antes de agregar los nuevos
+        if ($user->roles()->exists()) {
+            $user->roles()->detach();
+        }
+
+        if(!empty(request()->roles)) {
+            foreach (request()->roles as $role) {
+                $user->roles()->syncWithoutDetaching([
+                    $role => [
+                        'model_type' => get_class($user),
+                        'team_id' => $user->id
+                    ]
+                ]);
+            }
+        }
+
         return redirect('dashboard')->with(['success' => 'Usuario creado exitosamente.']);
     }
 
@@ -41,8 +63,9 @@ class InfoUserController extends Controller
 
     public function update(Request $request)
     {
+        // Retrieve the user model instance
+        $user = User::find(Auth::user()->id);
 
-        $user = User::where('id', Auth::user()->id);
         if ($_POST && $user) {
             $attributes = request()->validate([
                 'name' => ['required', 'max:50'],
@@ -52,8 +75,9 @@ class InfoUserController extends Controller
                 'about_me'    => ['max:150'],
                 'profile_picture' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
             ]);
+
             if ($request->get('email') != Auth::user()->email) {
-                //Sección para el rol
+                // Sección para el rol
                 if (env('IS_DEMO') && Auth::user()->id == 1) {
                     return redirect()->back()->withErrors(['msg2' => 'You are in a demo version, you can\'t change the email address.']);
                 }
@@ -72,22 +96,28 @@ class InfoUserController extends Controller
             ];
 
             if (request()->hasFile('profile_picture')) {
-                $dataUpdate['profile_picture'] = request()->file('profile_picture')->store('profile_pictures', 'public');
+                $profilePicture = ImageUploadController::upload($request->file('profile_picture'), 'profile_pictures', Auth::user()->name);
+                $dataUpdate['profile_picture'] = !empty($profilePicture) ? $profilePicture : Auth::user()->profile_picture;
             }
 
             $user->update($dataUpdate);
-            // Eliminar roles previos antes de agregar los nuevos
-            $user->roles()->detach();
 
-            $rolesIds = $request->roles;
-            foreach ($rolesIds as $role) {
-                $user->roles()->syncWithoutDetaching([
-                    $role => [
-                        'model_type' => get_class($user),
-                        'team_id' => $user->id
-                    ]
-                ]);
+            // Eliminar roles previos antes de agregar los nuevos
+            if ($user->roles()->exists()) {
+                $user->roles()->detach();
             }
+
+            if(!empty($request->roles)) {
+                foreach ($request->roles as $role) {
+                    $user->roles()->syncWithoutDetaching([
+                        $role => [
+                            'model_type' => get_class($user),
+                            'team_id' => $user->id
+                        ]
+                    ]);
+                }
+            }
+
         }
 
         return redirect('/user-profile')->with('success', '¡Perfil actualizado Correctamente!');
